--- InterviewSys_2/src/LoginPage/Login.jsx
@@ -1,32 +1,0 @@
-import React, { useState } from 'react';
-import './Login.css'; // Make sure to create and import your CSS file
-
-const Login = () => {
-    const [username, setUsername] = useState('');
-
-    const handleLogin = (e) => {
-        e.preventDefault();
-        console.log('Username:', username);
-    };
-
-    return (
-        <div className="login-container">
-            <h1>Login</h1>
-            <form className="login-form" onSubmit={handleLogin}>
-                <div className="form-group">
-                    <label htmlFor="username">Username</label>
-                    <input
-                        type="text"
-                        id="username"
-                        value={username}
-                        onChange={(e) => setUsername(e.target.value)}
-                        required
-                    />
-                </div>
-                <button type="submit" className="login-button">Login</button>
-            </form>
-        </div>
-    );
-};
-
-export default Login;
