+++ InterviewSys_2/src/main.jsx
@@ -5,11 +5,16 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import EnterScore from "./pages/EnterScore/EnterScore.tsx";
 import DefaultPage from "./defaultPage.jsx";
 import Table from "./pages/Table/table.jsx";
+import LoginPage from "./pages/LoginPage/Login.jsx";
 
 const router = createBrowserRouter([
     {
         path: "/",
         element: <DefaultPage />,
+    },
+    {
+        path: "/login",
+        element: <LoginPage />,
     },
     {
         path: "/table",
