+++ InterviewSys_2/src/pages/LoginPage/Login.css
@@ -4,7 +4,9 @@
     align-items: center;
     justify-content: center;
     height: 100vh;
-    background-color: #f5f5f5;
+    background-image: url('./screenshot.jpg');
+    background-size: cover;
+    background-position: center;
 }
 
 .login-container h1 {
